feat(backend): separate deploy and feeding logics
